pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Agentic RAG Chatbot with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchainhub>=0.1.21",
    "langsmith>=0.1.0",
    "pypdf>=5.0.0",
    "pydantic>=2.0.0",
    "gradio>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.0.0",
]
